Structure and configuration fixes
