+++ test_samples/weather-agent-framework/.env.example
@@ -1,3 +1,8 @@
+# M365 Agents SDK configuration
+CONNECTIONS__SERVICE_CONNECTION__SETTINGS__CLIENTID=client-id-here
+CONNECTIONS__SERVICE_CONNECTION__SETTINGS__CLIENTSECRET=client-secret-here
+CONNECTIONS__SERVICE_CONNECTION__SETTINGS__TENANTID=tenant-id-here
+
 # Azure OpenAI Configuration
 AZURE_OPENAI_ENDPOINT=https://your-resource.openai.azure.com/
 AZURE_OPENAI_API_KEY=your-api-key-here
@@ -8,6 +13,7 @@ AZURE_OPENAI_API_VERSION=2024-12-01-preview
 # Get your free API key from: https://openweathermap.org/price
 OPENWEATHER_API_KEY=your-openweather-api-key
 
-# Server Configuration
-HOST=localhost
-PORT=3978
+
+# Observability (optional)
+# OTEL_EXPORTER_OTLP_ENDPOINT=http://localhost:4317
+# APPLICATIONINSIGHTS_CONNECTION_STRING=InstrumentationKey=your-key-here
