+++ test_samples/weather-agent-framework/agents/__init__.py
@@ -1,0 +1,3 @@
+from .weather_agent import WeatherAgent
+
+__all__ = ["WeatherAgent"]
